Add friends likes

+++ src/main/java/ru/yandex/practicum/filmorate/model/Model.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public abstract class Model {
 
-    protected int id;
+    protected Long id;
 
 }
